pyproject.toml
[tool.poetry]
name = "fastapi-quiz"
version = "0.1.0"
description = ""
authors = ["alina1412 <alina1412@mail.ru>"]
readme = "README.md"
packages = [{include = "fastapi_quiz"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.19.0"
fastapi = "^0.87.0"
python-dotenv = "^0.21.0"
alembic = "^1.13.2"
black = "^24.8.0"
pylint = "^3.2.6"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
